Fix errors occurred with Intervention Image 3.3.1
In the course of the Intervention Image update, there was a
change which means that it is no longer possible to
instantiate so-called "driver-specific" modifiers yourself.
In this package, this initialization is always done by a
generic class via the respective driver class. This fix
sets the initialization to the static helper method used
by the driver of the dependency itself.