Fix TextBox line height type (#34)
* Fix TextBox line height type

\Intervention\Image\Typography\FontFactory lets you call `lineHeight`
with a float for more line height options. Currently, the package
only lets you provide a custom line height as an int.

Since the default value is 1.6, this fix lets you set it to e.g. 1.8,
which may be better if you are using a custom font or want to have
more granular control over the layout that is being rendered.

* Add PHP 8.4 to test matrix